fix: Hide feedback when offline
